v0.1.5
1. Simplify GUI, reduce number cached files and remove the concept of "project root".
2. Upgrade `numpy` from 1.19 to 1.24. Fix the problem due to the deletion of data type `np.int` etc.
3. Upgrade `paddlepaddle`.
4. Remove the "markup" function in single and clipboard recognition.

+++ text_recognition.py
@@ -109,7 +109,7 @@ def recognize_text(image, boxes, text_thresh=0.5, angle_classification_thresh=0.
             rec_res_final.append({
                 'text': text_,
                 'confidence': float(score),
-                'text_box_position': boxes[index_].astype(np.int).tolist()
+                'text_box_position': boxes[index_].astype(np.int32).tolist()
             })
     # List[ each_block: Dict["text": str, "confidence": float, "text_box_position": List[4*2] ] ]
     # text_box_position: [LU: [col_id, row_id], RU: ..., RD: ..., LD: ...]
